requirements.txt
asgiref==3.7.2
Django==4.2.7
django-admin-interface==0.28.5
django-colorfield==0.11.0
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
Markdown==3.5.2
pillow==10.2.0
psycopg2==2.9.9
python-slugify==8.0.4
pytz==2024.1
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2024.1
